Create torres country visa
The Torres county visa center generates the token number for its applicants from their application ID. The application ID is a numeric value. The token number generated in a specific form. The even digits in the applicant’s ID are replaced by the digit one greater then the even digit and the odd digits in the applicant’s ID are replaced by the digit one less than the odd digit. The numeric value thus generated represents the taken number of the applicant.

Input Format

The input consists of an integer – applicantID, representing the application ID of the applicant.

Constraints

1<=n<=10^7

Output Format

Print an integer representing the token number.