Create trailing zeroes 12
You are playing an online game. In the game, a numbers is displayed on the screen. In order to win the game, you have to Count the trailing zeros in the factorial value of the given number. Write an algorithm to count the trailing zeros in the factorial value of the given number.

Input Format

The input consists of an integer num, representing the number displayed on the screen.

Constraints

1<=n<=10^7

Output Format

Print An integer representing the count of trailing zeros in the factorial of the given numbers.